requirements.txt
google-adk>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
google-cloud-aiplatform>=1.38.0
google-genai>=0.6.0
python-dotenv>=1.0.0
pytest>=8.0.0